% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSize.R
\name{file.sizeWOLinks}
\alias{file.sizeWOLinks}
\alias{file.sizeCompare}
\title{Calculate file sizes omitting hard links}
\usage{
file.sizeWOLinks(path = ".", units = "auto", recursive = FALSE)

file.sizeCompare(path = ".", units = "auto", recursive = TRUE)
}
\arguments{
\item{path}{The path to evaluate file sizes in. Must be a directory.}

\item{units}{the units to be used in formatting and printing the size.
    Allowed values for the different \code{standard}s are
    \describe{
      \item{\code{standard = "legacy"}:}{
	\code{"b"}, \code{"Kb"}, \code{"Mb"}, \code{"Gb"}, \code{"Tb"}, \code{"Pb"},
	\code{"B"}, \code{"KB"}, \code{"MB"}, \code{"GB"}, \code{"TB"} and \code{"PB"}.}
      \item{\code{standard = "IEC"}:}{
	\code{"B"}, \code{"KiB"}, \code{"MiB"}, \code{"GiB"},
	\code{"TiB"}, \code{"PiB"}, \code{"EiB"}, \code{"ZiB"} and \code{"YiB"}.}
      \item{\code{standard = "SI"}:}{
	\code{"B"}, \code{"kB"}, \code{"MB"}, \code{"GB"}, \code{"TB"}, \code{"PB"},
	\code{"EB"}, \code{"ZB"}, \code{"YB"}, \code{"RB"}, and \code{"QB"}.}
    }
    For all standards, \code{unit = "auto"} is also allowed.
    If \code{standard = "auto"}, any of the "legacy" and \acronym{IEC}
    units are allowed.
    See \sQuote{Formatting and printing object sizes} for details.}

\item{recursive}{Logical indicating whether to search recursively.}
}
\description{
If two files are hardlinks, they don't actually take up extra space
on disk: there is only one copy of the data and two pointers to
the data.
}
