% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedev-package.R
\docType{package}
\name{pedev-package}
\alias{pedev}
\alias{pedev-package}
\title{pedev}
\description{
Miscellaneous development tools developed by the Predictive Ecology Group (\url{https://predictiveecology.org}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedev.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/pedev}
  \item Report bugs at \url{https://github.com/PredictiveEcology/pedev/issues}
}

}
\author{
\strong{Maintainer}: Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
