% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDotUnderline.R
\name{rmDotUnderline}
\alias{rmDotUnderline}
\title{Remove objects that start with "dot" "underscore"}
\usage{
rmDotUnderline(envir = .GlobalEnv)
}
\arguments{
\item{envir}{The environment from with to rm objects, defaults to
\code{.GlobalEnv}}
}
\description{
A quick wrapper for removing a objects with specials names,
\code{"._xxx"}.
}
