% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTools.R
\name{reload_all}
\alias{reload_all}
\title{A verion of devtools::load_all that detaches dependencies}
\usage{
reload_all(pkgs, load_all = TRUE, gitPath = "~/GitHub")
}
\arguments{
\item{pkgs}{A character vector of the package(s) to run "devtools::load_all"}

\item{load_all}{Logical. If \code{FALSE}, then this function will only
detach the packages necessary}
}
\description{
This is very idiosyncratic for the Predictive Ecology group
}
