% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTools.R
\name{updateGit}
\alias{updateGit}
\title{Pull from git and install}
\usage{
updateGit(pkgs = NULL, install = TRUE, branch = "development")
}
\arguments{
\item{pkgs}{A character vector of package names, which is actually
the path names of the packages. i.e., must be absolute or relative
path. Defaults to current directory. It}

\item{install}{Logical. If TRUE, then it will run devtools::install if
there is new content}

\item{branch}{The branch to pull from.}
}
\description{
Fetches all branches, then pulls the identified branch from git,
then runs a digest on the local folders. If that digest is different
as a previous one, then the function will run \code{devtools::install}.
}
